Add functions to get values and calculate a winner
Functions get input either provided randomly or by a prompt to return
their values. added switch/case statement that calculates the winner,
although it is not inside of a function and there's no global variable
to store player's or computer's choice, so through the conditional that
determines the winners, the values are calculate once again.