Add playGame function to run the game five times
The function is called once, and runs playRound five
times while keeping track of both computer's and player's
scores. Adds a function called showResults, that is declared
and ran inside of playGame, and it shows who the winner is
after the rounds are run. Variables to keep track of scores
have been moved to inside of playGame.