Add buttons to set player choice
The buttons are selected with querySelector on a UL and handled
with a single event listener, which consumes less memory than
if there were event listeners for each button. The playGame function
no longer exists. Computer choice, round winner and results of
each game are still shown on the console.