Remove stacking blocks of text on the page
The results of each round and match would be displayed
below the existing text, which could potentially cause
the page to be overwhelmed with text or the user to be
confused about the results.

Now, each round's results replace the previous one. The
game's results, which triggers when someone reach 3 points,
also follows this same logic.